+++ edas-20170801/src/Models/DescribeLocalitySettingRequest.php
@@ -1,0 +1,76 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Edas\V20170801\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class DescribeLocalitySettingRequest extends Model
+{
+    /**
+     * @var string
+     */
+    public $appId;
+
+    /**
+     * @var string
+     */
+    public $namespaceId;
+
+    /**
+     * @var string
+     */
+    public $region;
+    protected $_name = [
+        'appId' => 'AppId',
+        'namespaceId' => 'NamespaceId',
+        'region' => 'Region',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->appId) {
+            $res['AppId'] = $this->appId;
+        }
+
+        if (null !== $this->namespaceId) {
+            $res['NamespaceId'] = $this->namespaceId;
+        }
+
+        if (null !== $this->region) {
+            $res['Region'] = $this->region;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['AppId'])) {
+            $model->appId = $map['AppId'];
+        }
+
+        if (isset($map['NamespaceId'])) {
+            $model->namespaceId = $map['NamespaceId'];
+        }
+
+        if (isset($map['Region'])) {
+            $model->region = $map['Region'];
+        }
+
+        return $model;
+    }
+}