+++ edas-20170801/src/Models/DescribeLocalitySettingResponseBody/data.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Edas\V20170801\Models\DescribeLocalitySettingResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class data extends Model
+{
+    /**
+     * @var bool
+     */
+    public $enabled;
+
+    /**
+     * @var float
+     */
+    public $threshold;
+    protected $_name = [
+        'enabled' => 'Enabled',
+        'threshold' => 'Threshold',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->enabled) {
+            $res['Enabled'] = $this->enabled;
+        }
+
+        if (null !== $this->threshold) {
+            $res['Threshold'] = $this->threshold;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Enabled'])) {
+            $model->enabled = $map['Enabled'];
+        }
+
+        if (isset($map['Threshold'])) {
+            $model->threshold = $map['Threshold'];
+        }
+
+        return $model;
+    }
+}