+++ edas-20170801/ChangeLog.md
@@ -1,3 +1,10 @@
+2025-11-03 Version: 1.1.0
+- Support API DescribeLocalitySetting.
+- Update API DeployK8sApplication: add request parameters SecurityContext.
+- Update API GetK8sApplication: add response parameters Body.Applcation.App.SecurityContext.
+- Update API InsertK8sApplication: add request parameters SecurityContext.
+
+
 2025-10-28 Version: 1.0.9
 - Generated php 2017-08-01 for Edas.